Add files via upload
A fun and small interactive app using HTML - CSS - JavaScript, which stores groceries (and other to-dos). The user can remove any entry in the list and clear the whole list well. There is also a filter bar available, where the user can search for any entry. The list gets saved in the local storage so the user can find the list with the entries any time.